Move pruning function to be called before json export is created
